+++ spec/app_info_spec.rb
@@ -6,6 +6,7 @@ MATCH_FILE_TYPES = {
   'android.apk' => :apk,
   'tv.apk' => :apk,
   'wear.apk' => :apk,
+  'android.aab' => :aab,
   'ipad.ipa' => :ipa,
   'iphone.ipa' => :ipa,
   'embedded.ipa' => :ipa,
@@ -22,6 +23,8 @@ MATCH_FILE_TYPES = {
   'macos_development.provisionprofile' => :mobileprovision,
   'single_mapping.zip' => :proguard,
   'full_mapping.zip' => :proguard,
+  'macos.zip' => :macos,
+  'macos-signed.zip' => :macos
 }
 
 describe AppInfo do
@@ -54,6 +57,8 @@ describe AppInfo do
             expect(parse).to be_a(AppInfo::DSYM)
           when :mobileprovision
             expect(parse).to be_a(AppInfo::MobileProvision)
+          when :macos
+            expect(parse).to be_a(AppInfo::Macos)
           end
         end
       end